# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PoplarGrid API 服务器，提供项目、成员等操作功能",
        "title": "PoplarGrid 核心 API 服务器",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/worksets": {
            "get": {
                "description": "注意当列表为空，会返回 null 而不是空数组",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workset"
                ],
                "summary": "获取特定汉化组的工作集列表分页，按 ID 倒序",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码，默认值为 1",
                        "name": "page_serial",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认值为 10",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "所属汉化组 ID",
                        "name": "team_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.WorksetBasic"
                            }
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/worksets/{id}/stats": {
            "get": {
                "description": "获取所有项目的统计信息，包括总数、翻译、校对、嵌字，审核、发布对应数量等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workset"
                ],
                "summary": "获取特定作品集项目统计信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "作品集 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ProjectStats"
                        }
                    },
                    "400": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dtos.ProjectStats": {
            "type": "object",
            "properties": {
                "lettered_count": {
                    "description": "已嵌字的项目数量",
                    "type": "integer"
                },
                "lettering_count": {
                    "description": "正在嵌字的项目数量",
                    "type": "integer"
                },
                "not_lettering_count": {
                    "description": "未开始嵌字的项目数量",
                    "type": "integer"
                },
                "not_prooving_count": {
                    "description": "未开始校对的项目数量",
                    "type": "integer"
                },
                "not_reviewing_count": {
                    "description": "未开始审核的项目数量",
                    "type": "integer"
                },
                "not_translating_count": {
                    "description": "未开始翻译的项目数量",
                    "type": "integer"
                },
                "prooved_count": {
                    "description": "已校对的项目数量",
                    "type": "integer"
                },
                "prooving_count": {
                    "description": "正在校对的项目数量",
                    "type": "integer"
                },
                "published_count": {
                    "description": "已发布项目数量",
                    "type": "integer"
                },
                "reviewed_count": {
                    "description": "已审核的项目数量",
                    "type": "integer"
                },
                "reviewing_count": {
                    "description": "正在审核的项目数量",
                    "type": "integer"
                },
                "total_count": {
                    "description": "总项目数量",
                    "type": "integer"
                },
                "translated_count": {
                    "description": "已翻译的项目数量",
                    "type": "integer"
                },
                "translating_count": {
                    "description": "正在翻译的项目数量",
                    "type": "integer"
                },
                "workset_id": {
                    "description": "对应作品集 ID",
                    "type": "integer"
                }
            }
        },
...
        "dtos.WorksetBasic": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "作品集 ID",
                    "type": "integer"
                },
                "name": {
                    "description": "作品集名称",
                    "type": "string"
                },
                "team_id": {
                    "description": "所属团队 ID",
                    "type": "integer"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "可选，在 service 层发生错误时提供详细信息",
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}